+++ vue-server/src/demo/integrations/router/browser.ts
@@ -1,0 +1,25 @@
+export function listenHistory(onNavigation: () => void) {
+	window.addEventListener("pushstate", onNavigation);
+	window.addEventListener("popstate", onNavigation);
+
+	const oldPushState = window.history.pushState;
+	window.history.pushState = function (...args) {
+		const res = oldPushState.apply(this, args);
+		onNavigation();
+		return res;
+	};
+
+	const oldReplaceState = window.history.replaceState;
+	window.history.replaceState = function (...args) {
+		const res = oldReplaceState.apply(this, args);
+		onNavigation();
+		return res;
+	};
+
+	return () => {
+		window.removeEventListener("pushstate", onNavigation);
+		window.removeEventListener("popstate", onNavigation);
+		window.history.pushState = oldPushState;
+		window.history.replaceState = oldReplaceState;
+	};
+}
