+++ vue-server/src/demo/routes/_client.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { tinyassert } from "@hiogawa/utils";
 import { defineComponent, inject, onMounted, ref } from "vue";
 
 export const ClientCounter = defineComponent(() => {
@@ -28,62 +27,6 @@ export const ClientNested = defineComponent((_props, { slots }) => {
 		</span>
 	);
 });
-
-export const Link = defineComponent<{ href: string }>(
-	(props, { slots }) => {
-		return () => (
-			<a
-				href={props.href}
-				onClick={(e) => {
-					if (
-						e.currentTarget instanceof HTMLAnchorElement &&
-						e.button === 0 &&
-						!(e.metaKey || e.altKey || e.ctrlKey || e.shiftKey) &&
-						(!e.currentTarget.target || e.currentTarget.target === "_self")
-					) {
-						e.preventDefault();
-						history.pushState(null, "", e.currentTarget.href);
-					}
-				}}
-			>
-				{slots.default?.()}
-			</a>
-		);
-	},
-	{
-		props: ["href"],
-	},
-);
-
-export const Form = defineComponent<{ replace?: boolean }>(
-	(props, { slots }) => {
-		return () => (
-			<form
-				onSubmit={(e) => {
-					e.preventDefault();
-					tinyassert(e.currentTarget instanceof HTMLFormElement);
-					const url = new URL(e.currentTarget.action);
-					const data = new FormData(e.currentTarget);
-					data.forEach((v, k) => {
-						if (typeof v === "string") {
-							url.searchParams.set(k, v);
-						}
-					});
-					if (props.replace) {
-						history.replaceState({}, "", url);
-					} else {
-						history.pushState({}, "", url);
-					}
-				}}
-			>
-				{slots.default?.()}
-			</form>
-		);
-	},
-	{
-		props: ["replace"],
-	},
-);
 
 export const GlobalProgress = defineComponent(() => {
 	const isLoading = inject("isLoading", { value: false });
